some updates

+++ db/seeds.rb
@@ -10,7 +10,7 @@ require 'json'
 require 'faker'
 
 ########### reset all ########
-
+=begin
 if Image
   Item.image.destroy_all
 end
@@ -20,12 +20,12 @@ end
 if User
   User.destroy_all
 end
-/if category
+if category
   category.destroy_all
 end
-/
+=end
 ############ Seed des images
-
+=begin
 Image.create(
   title: "boitier nikon d300s",
   img_url: "app/assets/images/boitiers/boitier-nikon-d300s.jpg"
@@ -41,18 +41,19 @@ Item.create!(
   user_id: "1",
   category_id: "1"
   )
-
+=end
 
 ###########  Seed Users
-User.create!(
-  email: bar@foo.fr,
+User.create!([ {
+  email: "bar@foo.fr",
   postal_code: "92500",
   user_type: "owner", #can have two values: owner or customer
   first_name: "Rado",
   last_name: "Rakoto",
   about: "amateur photo depuis quelques années, je vous propose mon matériel à la location",
   img_cin_url: ""
-)
+}
+  ]  )
 
 ########### Seed Categories
 #todo: creation model
